[GStreamer][EME] Prevent race conditions when flushing https://bugs.webkit.org/show_bug.cgi?id=246682
Reviewed by Philippe Normand.

In some cases, decryption can be slower than flush processing and we might be setting and unsetting the flushing flag
without the decryption thread noticing. We add a small state machine to ensure that if we get a flush while decrypting,
we are going to ditch that buffer even if the result was correctly decrypted.

* Source/WebCore/platform/graphics/gstreamer/eme/WebKitCommonEncryptionDecryptorGStreamer.cpp:
(transformInPlace):
(sinkEventHandler):

Canonical link: https://commits.webkit.org/255673@main